Add Phase 1a project scaffold, services, and unit tests

+++ src/NdiTelop/Models/BackgroundStyle.cs
@@ -1,8 +1,7 @@
 namespace NdiTelop.Models;
-
 public class BackgroundStyle
 {
     public string Type { get; set; } = "transparent";
     public string Color { get; set; } = "#000000";
-    public double Alpha { get; set; }
+    public float Alpha { get; set; }
 }
